+++ include/check_network_status.h
@@ -1,0 +1,4 @@
+#ifndef CHECKNETWORKSTATUS_H
+#define CHECKNETWORKSTATUS_H
+#include "../src/ui/check_network_status.c"
+#endif
