+++ src/ui/createMainWindow.c
@@ -1,26 +1,6 @@
-#include "./onStartClicked.h"
-#include "createButton.h"
+#include "check_network_status.h"
 #include "createImage.h"
-
-gboolean check_network_status(gpointer user_data) {
-
-  NetworkStatusData *app_data = (NetworkStatusData *)user_data;
-
-  g_print("Running check network status.\n");
-  if (findMyIP().status != -1) {
-    gtk_label_set_text(GTK_LABEL(app_data->label),
-                       "You are ready to start the SSDP Listener. Please click "
-                       "the 'Start' button below.");
-    gtk_widget_set_sensitive(app_data->button, TRUE);
-  } else {
-    gtk_label_set_text(
-        GTK_LABEL(app_data->label),
-        "Your device needs to be connected to any network before you proceed. "
-        "Please check the network connection of your device");
-    gtk_widget_set_sensitive(app_data->button, FALSE);
-  }
-  return G_SOURCE_CONTINUE;
-}
+#include "onStartClicked.h"
 
 // Create a main window i.e. first window to show when application launches
 GtkWidget *create_main_window(GtkApplication *app) {
@@ -36,7 +16,7 @@ GtkWidget *create_main_window(GtkApplication *app) {
   GtkWidget *logo_image = create_image("../include/image.png", 200, 200);
 
   start_data->timeout_id =
-      g_timeout_add_seconds(4, check_network_status, app_data);
+      g_timeout_add_seconds(5, check_network_status, app_data);
 
   GtkWidget *start_button =
       create_button("Start", G_CALLBACK(on_start_clicked), start_data);
