+++ src/ui/onStartClicked.c
@@ -5,11 +5,11 @@
 
 void on_start_clicked(GtkButton *button, gpointer data) {
   StartData *user_data = (StartData *)data;
-  GtkWindow *window =
-      GTK_WINDOW(gtk_widget_get_ancestor(GTK_WIDGET(button), GTK_TYPE_WINDOW));
-
   UiUpdateData *ui_data = g_new0(UiUpdateData, 1);
   JoinSSDPThread *thread_data = g_new0(JoinSSDPThread, 1);
+
+  GtkWindow *window =
+      GTK_WINDOW(gtk_widget_get_ancestor(GTK_WIDGET(button), GTK_TYPE_WINDOW));
 
   if (user_data->timeout_id > 0) {
     g_source_remove(user_data->timeout_id);
@@ -24,9 +24,11 @@ void on_start_clicked(GtkButton *button, gpointer data) {
   GtkWidget *spinner = gtk_spinner_new();
   gtk_widget_set_size_request(GTK_WIDGET(spinner), 50, 50);
   gtk_spinner_start(GTK_SPINNER(spinner));
+  gtk_box_append(GTK_BOX(vertical_box_for_spinner), spinner);
 
   GtkWidget *spin_label = create_wrapped_label("Scanning for devices...", 30);
   gtk_widget_set_margin_top(spin_label, 20);
+  gtk_box_append(GTK_BOX(vertical_box_for_spinner), spin_label);
 
   GTask *task = g_task_new(NULL, NULL, on_ssdp_scan_completed, ui_data);
   g_task_run_in_thread(task, run_ssdp_scan);
@@ -37,12 +39,7 @@ void on_start_clicked(GtkButton *button, gpointer data) {
   GThread *join_ssdp_thread = g_thread_new("join_ssdp", joinSSDP, thread_data);
   g_thread_unref(join_ssdp_thread);
 
-  gtk_box_append(GTK_BOX(vertical_box_for_spinner), spinner);
-  gtk_box_append(GTK_BOX(vertical_box_for_spinner), spin_label);
-
   gtk_window_set_child(window, vertical_box_for_spinner);
 
-  // ui_data->spinner = spinner;
-  // ui_data->spin_label = spin_label;
   ui_data->window = window;
 }
